@@ -1,5 +1,9 @@
 import { bootstrapApplication } from '@angular/platform-browser';
 import { appConfig } from './app.config';
 import { AppComponent } from './app.component';
+import { register } from 'swiper/element/bundle';
 
-bootstrapApplication(AppComponent, appConfig).catch((err) => console.error(err));
+register();
+bootstrapApplication(AppComponent, appConfig).catch((err) =>
+  console.error(err),
+);
